some cleanup to improve overall structure
