+++ .env.emulators
@@ -1,0 +1,1 @@
+RECIPE_HANDLER_USE_EMULATORS=true
