+++ src/Components/ErrorBoundary.tsx
@@ -44,7 +44,7 @@ class ErrorBoundary extends Component<{}, ErrorBoundaryState> {
     FirebaseService.firestore
       .collection('errors')
       .add({
-        version: APP_VERSION,
+        version: RECIPE_HANDLER_APP_VERSION,
         minError: minError.toString(),
         trace: trace.toString(),
         agent: window.navigator.userAgent,
