+++ src/vite-env.d.ts
@@ -2,11 +2,12 @@
 /// <reference types="vite-plugin-svgr/client" />
 /// <reference types="vite-plugin-pwa/client" />
 
-interface ImportMetaEnv {}
+interface ImportMetaEnv {
+  RECIPE_HANDLER_USE_EMULATORS: boolean
+}
 
 interface ImportMeta {
   readonly env: ImportMetaEnv
 }
 
-declare const APP_VERSION: string
-declare const USE_EMULATORS: boolean
+declare const RECIPE_HANDLER_APP_VERSION: string
